pyproject.toml
[tool.poetry]
name = "ten-drops"
version = "0.1.0"
description = "A ten drops game written in pygame."
authors = ["chyok <chyok@hotmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "ten_drops"}]

[tool.poetry.dependencies]
python = "^3.11"
pygame-ce = "^2.4.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
